refactor: move popover to src folder and fix reactRef propType
