+++ src/prop-validators/reactRef.js
@@ -1,5 +1,5 @@
 import propTypes from 'prop-types'
 
 export const reactRef = propTypes.shape({
-    current: propTypes.element,
+    current: propTypes.node,
 })
